+++ src/test/java/racing/CarTest.java
@@ -3,7 +3,8 @@ package racing;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.Car;
+import step3.domain.Car;
+import step3.domain.DecideMode;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -21,14 +22,14 @@ public class CarTest {
     @Test
     void CarForWardTrueModeTest() {
         Car car = new Car();
-        car.move(Car.DecideMode.TRUE_RETURN_MODE);
+        car.move(DecideMode.TRUE_RETURN_MODE);
         assertThat(car.getDistance()).isEqualTo(1);
     }
 
     @Test
     void CarForWardFalseModeTest() {
         Car car = new Car();
-        car.move(Car.DecideMode.FALSE_RETURN_MODE);
+        car.move(DecideMode.FALSE_RETURN_MODE);
         assertThat(car.getDistance()).isEqualTo(0);
     }
