+++ src/test/java/racing/RacingGameTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
-import step3.RacingGame;
+import step3.domain.RacingGame;
 
 import java.util.stream.Stream;
